refactor: auth utils moved to a separate file
